Create InertiaTable vue file

+++ src/Commands/MakeInertiaTable.php
@@ -2,7 +2,11 @@
 
 namespace harmonic\InertiaTable\Commands;
 
+use Symfony\Component\Process\Process;
+use Symfony\Component\Process\Exception\ProcessFailedException;
 use Illuminate\Support\Str;
+use File;
+use Illuminate\Support\Facades\Schema;
 use Illuminate\Console\Command;
 
 class MakeInertiaTable extends Command
@@ -33,6 +37,26 @@ class MakeInertiaTable extends Command
         parent::__construct();
     }
 
+    private function inertiaTableJSExists() {
+        $process = new Process('npm list --depth=0 | grep inertia-table');
+        $process->run();
+        
+        // executes after the command finishes
+        if (!$process->isSuccessful()) {
+            if ($process->getOutput()=="") {
+                return false;
+            }
+            $this->error('Failed looking for inertia-table JS package. Is inertia installed?');
+            throw new ProcessFailedException($process);
+        }
+
+        $result = $process->getOutput();
+        if (strpos($result, ' inertia-table@') !== false) {
+            return true;
+        }
+        return false;
+    }
+
     /**
      * Execute the console command.
      *
@@ -40,26 +64,47 @@ class MakeInertiaTable extends Command
      */
     public function handle()
     {
-        $model = $this->argument('model');
+        $model = ucfirst($this->argument('model'));
 
-        // Create Model
-        $str = file_get_contents(__DIR__ . '/../Stubs/Model.php');
-        $str = str_replace("@namespace", substr($this->getAppNamespace(), 0, -1), $str);
-        $str = str_replace("@modelName", $model, $str);
-        file_put_contents(app_path($model . '.php'), $str);
+        if (!file_exists(app_path($model).'.php') && !file_exists(app_path('Models\\' . $model).'.php')) {
+            // Create Model
+            $str = file_get_contents(__DIR__ . '/../Stubs/Model.php');
+            $str = str_replace("@namespace", substr($this->getAppNamespace(), 0, -1), $str);
+            $str = str_replace("@modelName", $model, $str);
+            file_put_contents(app_path($model . '.php'), $str);
+        }
 
         // Create Controller
-        $controllerName = Str::plural($model) . 'Controller';
+        $pluralName = Str::plural($model);
+        $controllerName = $pluralName . 'Controller';
         $str = file_get_contents(__DIR__ . '/../Stubs/Controller.php');
         $str = str_replace("@namespace", substr($this->getAppNamespace(), 0, -1), $str);
         $str = str_replace("@controllerName", $controllerName, $str);
         $str = str_replace("@modelName", $model, $str);
         file_put_contents(app_path('Http/Controllers/' . $controllerName . '.php'), $str);
 
-        //TODO: Add routes
+        // Add routes
         $this->warn('You need to manually add route: Route::inertia(\''. Str::plural(strtolower($model)) . '\');');
 
-        //TODO: Create a InertiaTable.vue component and then create Index.vue using it
+        if ($this->inertiaTableJSExists()) {
+            $pluralNameLowercase = \strtolower($pluralName);
+            $modelLowercase = \strtolower($model);
+
+            // Get column names from table
+            $class = 'App\\' . $model;
+            $columns = Schema::getColumnListing(with(new $class)->getTable());
+            
+            // Create an InertiaTable.vue component and then create Index.vue using it
+            $tableElementPath = resource_path('js/Pages/' . $pluralName);
+            File::makeDirectory($tableElementPath, 0755, true);
+            $str = file_get_contents(__DIR__ . '/../Stubs/Index.vue');
+            $str = str_replace("@pluralNameLowercase", $pluralNameLowercase, $str);
+            $str = str_replace("@plural", $pluralName, $str);
+            $str = str_replace("@columns", implode(",", $columns), $str);
+            $str = str_replace("@modelLowercase", $modelLowercase, $str);
+            file_put_contents($tableElementPath . '/Index.vue', $str);
+        }
+
         $this->info('Model, Controller and Vue Components successfully created.');
     }
 }
